Move version check code out of class Updater
