+++ apps/updatenotification/appinfo/app.php
@@ -20,10 +20,9 @@
  */
 
 if(\OC::$server->getConfig()->getSystemValue('updatechecker', true) === true) {
-	$updater = new \OC\Updater(
+	$updater = new \OC\Updater\VersionCheck(
 		\OC::$server->getHTTPHelper(),
-		\OC::$server->getConfig(),
-		\OC::$server->getIntegrityCodeChecker()
+		\OC::$server->getConfig()
 	);
 	$updateChecker = new \OCA\UpdateNotification\UpdateChecker(
 		$updater
